Compiler warning for 64 bit build

+++ Nocilla/Hooks/ASIHTTPRequest/ASIHTTPRequestStub.h
@@ -1,5 +1,5 @@
 @interface ASIHTTPRequestStub : NSObject
-- (int)stub_responseStatusCode;
+- (NSInteger)stub_responseStatusCode;
 - (NSData *)stub_responseData;
 - (NSDictionary *)stub_responseHeaders;
 - (void)stub_startRequest;
